requirements.txt
mss
opencv-python
pyautogui
stockfish
joblib
python-chess
numpy

